+++ docs/release_notes.md
@@ -9,6 +9,7 @@
 *   It is now possible to add arbitrary metadata to `tfds.core.DatasetInfo`
     which will be stored/restored with the dataset. See `tfds.core.Metadata`.
 *   Better proxy support, possibility to add certificate
+*   Add Speech Command dataset
 *   Add `decoders` kwargs to override the default feature decoding
     ([guide](https://github.com/tensorflow/datasets/tree/master/docs/decode.md)).
 *   Add `duke_ultrasound` dataset of ultrasound phantoms and invivo liver images